+++ backend/src/main/java/com/popcorn/demo/domain/payment/dto/response/OrderPaymentCreateResponse.java
@@ -1,6 +1,5 @@
-package com.popcorn.demo.domain.order.dto.response;
+package com.popcorn.demo.domain.payment.dto.response;
 
-import java.time.LocalDateTime;
 import java.util.UUID;
 
 import lombok.AllArgsConstructor;
@@ -12,10 +11,9 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class ReservationPaymentCreateResponse {
+public class OrderPaymentCreateResponse {
 
 	private UUID paymentId;
-	private String paymentStatus;
+	private String status;
 	private String orderStatus;
-	private LocalDateTime approvedAt;
 }
