+++ backend/src/main/java/com/popcorn/demo/domain/payment/entity/PaymentMethod.java
@@ -1,4 +1,4 @@
-package com.popcorn.demo.domain.order.entity;
+package com.popcorn.demo.domain.payment.entity;
 
 public enum PaymentMethod {
 	CARD,
